[core] pattern assignments now always return a list
Edited the pattern assignments so a list is always returned. This
however means that the compiler cannot handle assignments such as
```
a, *b, c = (1, True, 3, 4)
```
as lists in nac3 are homogenous. However assignments such as
```
x, *y, z = (True, 2, 3, 4)
```
are still supported as the elements in the tuple corresponding to the
starred target are homogenous.