[core] Moved static methods from a field on `Class` to function
Reverted `TopLevelDef::Class` back to its original state. Added a new
`attributes` field onto `TopLevelDef::Function`. This feels more correct
as static is a property of the function (which does nothing if the
function is not a class method). Additionally this is extensible in case
future attributes, e.g. `inline`, `async`, etc. are to be added.