Fix check for types requiring alloca
Fix the check for the type of a value returned by an RPC function.
Previous to this a simple `is_pointer_type()` check was being used,
which caused `stackrestore` to be called on types such as `tuple[list]`
as `tuple` is not a pointer type. This however, caused UB such as raised
in issue #577.