Added a function to fold over `BasicTypeEnum`s
Added `try_fold_basic_type`, a function that can recursively fold over a
`BasicTypeEnum` with support for short-circuiting given the input
function. This is accompanied by `basic_type_all` which checks that some
predicate holds for the full recursive structure of a `BasicTypeEnum`
and short circuits on a faliure of the predicate.