[artiq] Fix incorrect `stackrestore` causing corruption of some returned types (#653)
Fix the check for the type of a value returned by an RPC function. Previous to this a simple `is_pointer_type()` check was being used, which caused `stackrestore` to be called in too many instances. This caused UB, such as the corruption of a list returned in a tuple, as raised in issue #577 due to the tuple container not being a pointer type. To fix this, the datatype is now recursively checked to ensure that no part of it may have required further allocation, in which case the stack should not be restored.

Fixes #577.

Reviewed-on: https://git.m-labs.hk/M-Labs/nac3/pulls/653
Co-authored-by: Oliver Archer <oa@m-labs.hk>
Co-committed-by: Oliver Archer <oa@m-labs.hk>